backend/requirements.txt
action-msgs==2.0.3
action-tutorials-interfaces==0.33.10
action-tutorials-py==0.33.10
actionlib-msgs==5.3.7
ament-cmake-test==2.5.6
ament-copyright==0.17.5
ament-cppcheck==0.17.5
ament-cpplint==0.17.5
ament-flake8==0.17.5
ament-index-python==1.8.2
ament-lint==0.17.5
ament-lint-cmake==0.17.5
ament-package==0.16.5
ament-pep257==0.17.5
ament-uncrustify==0.17.5
ament-xmllint==0.17.5
angles==1.16.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
builtin-interfaces==2.0.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
composition-interfaces==2.0.3
cryptography==46.0.7
cv-bridge==4.1.0
demo-nodes-py==0.33.10
diagnostic-msgs==5.3.7
domain-coordinator==0.12.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.7
examples-rclpy-minimal-action-client==0.19.7
examples-rclpy-minimal-action-server==0.19.7
examples-rclpy-minimal-client==0.19.7
examples-rclpy-minimal-publisher==0.19.7
examples-rclpy-minimal-service==0.19.7
examples-rclpy-minimal-subscriber==0.19.7
fastapi==0.136.0
geometry-msgs==5.3.7
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.12
image-geometry==4.1.0
interactive-markers==2.5.5
laser-geometry==2.7.2
launch==3.4.10
launch-ros==0.26.11
launch-testing==3.4.10
launch-testing-ros==0.26.11
launch-xml==3.4.10
launch-yaml==3.4.10
lifecycle-msgs==2.0.3
logging-demo==0.33.10
lxml==6.1.0
map-msgs==2.4.1
message-filters==4.11.12
nav-msgs==5.3.7
oauthlib==3.3.1
ollama==0.6.1
osrf-pycommon==2.1.7
pcl-msgs==1.0.0
pendulum-msgs==0.33.10
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyparsing==3.3.2
PyPDF2==3.0.1
python-docx==1.2.0
python-qt-binding==2.2.2
qt-dotgraph==2.7.6
qt-gui==2.7.6
qt-gui-cpp==2.7.6
qt-gui-py-common==2.7.6
quality-of-service-demo-py==0.33.10
rcl-interfaces==2.0.3
rclpy==7.1.11
rcutils==6.7.5
requests==2.33.1
requests-oauthlib==2.0.0
resource-retriever==3.4.4
rmw-dds-common==3.1.1
ros2action==0.32.9
ros2bag==0.26.10
ros2bag-mcap-cli==0.26.10
ros2bag-sqlite3-cli==0.26.10
ros2cli==0.32.9
ros2component==0.32.9
ros2doctor==0.32.9
ros2interface==0.32.9
ros2launch==0.26.11
ros2lifecycle==0.32.9
ros2multicast==0.32.9
ros2node==0.32.9
ros2param==0.32.9
ros2pkg==0.32.9
ros2plugin==5.4.5
ros2run==0.32.9
ros2service==0.32.9
ros2topic==0.32.9
rosbag2-interfaces==0.26.10
rosbag2-py==0.26.10
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.7
rosidl-cli==4.6.7
rosidl-cmake==4.6.7
rosidl-generator-c==4.6.7
rosidl-generator-cpp==4.6.7
rosidl-generator-py==0.22.2
rosidl-generator-rs==0.4.11
rosidl-generator-type-description==4.6.7
rosidl-parser==4.6.7
rosidl-pycommon==4.6.7
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.3
rosidl-typesupport-fastrtps-cpp==3.6.3
rosidl-typesupport-introspection-c==4.6.7
rosidl-typesupport-introspection-cpp==4.6.7
rpyutils==0.4.2
rqt-action==2.2.1
rqt-bag==1.5.6
rqt-bag-plugins==1.5.6
rqt-console==2.2.2
rqt-graph==1.5.6
rqt-gui==1.6.3
rqt-gui-py==1.6.3
rqt-msg==1.5.2
rqt-plot==1.4.5
rqt-publisher==1.7.3
rqt-py-common==1.6.3
rqt-py-console==1.2.3
rqt-reconfigure==1.6.4
rqt-service-caller==1.2.2
rqt-shell==1.2.3
rqt-srv==1.2.3
rqt-topic==1.7.5
sensor-msgs==5.3.7
sensor-msgs-py==5.3.7
service-msgs==2.0.3
shape-msgs==5.3.7
sros2==0.13.6
starlette==1.0.0
statistics-msgs==2.0.3
std-msgs==5.3.7
std-srvs==5.3.7
stereo-msgs==5.3.7
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.36.20
tf2-kdl==0.36.20
tf2-msgs==0.36.20
tf2-py==0.36.20
tf2-ros-py==0.36.20
tf2-sensor-msgs==0.36.20
tf2-tools==0.36.20
topic-monitor==0.33.10
trajectory-msgs==5.3.7
turtlesim==1.8.3
type-description-interfaces==2.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.5.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.45.0
visualization-msgs==5.3.7
